codec: use record only
